apis/requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
google-auth==2.14.1
idna==3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==3.12.3
pytz==2022.6
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.12
